move content from file to config
