--- docs/lib/commandlist.ts
@@ -1,61 +1,0 @@
-export const links = [
-  {
-    name: 'Links',
-    items: [
-      {
-        name: 'Slack',
-        href: 'https://reservix.slack.com/archives/C02727BNZ3J',
-        keywords: ['help'],
-      },
-      {
-        name: 'Jira',
-        href: 'https://reservix.atlassian.net/jira/software/projects/DST/boards/134',
-      },
-      {
-        name: 'Figma Core Kit',
-        href: 'https://www.figma.com/design/NbTUW9zk15nN8idlfsEttS/%F0%9F%8C%BC-Marigold-CORE?t=VfTLYo5foFEjRxFY-0',
-        keywords: ['design'],
-      },
-      {
-        name: 'Support Center',
-        href: 'https://reservix.atlassian.net/servicedesk/customer/portal/77',
-        keywords: ['help'],
-      },
-      {
-        name: 'Storybook',
-        href: 'https://marigold-latest.vercel.app/?path=/story/components-accordion--basic',
-      },
-      {
-        name: 'Stackblitz',
-        href: 'https://stackblitz.com/edit/github-v1evcv?file=src%2FApp.tsx',
-        keywords: ['playground', 'starter'],
-      },
-      {
-        name: 'Reference App',
-        href: 'https://marigold-rapp.vercel.app/',
-        keywords: ['help', 'example', 'code'],
-      },
-      {
-        name: 'Github',
-        href: 'https://github.com/marigold-ui/marigold',
-        keywords: ['code'],
-      },
-    ],
-  },
-];
-
-export const themeswitch = [
-  {
-    name: 'Theme Switch',
-    items: [
-      {
-        name: 'Switch to Core Theme',
-        theme: 'core',
-      },
-      {
-        name: 'Switch to B2B Theme',
-        theme: 'b2b',
-      },
-    ],
-  },
-];
